+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Add ability to use a different username instead of hardcoding to 'winesap'.
 - Add support for major upgrades from Mac Linux Gaming Stick 2 to winesapOS 3.
 - Add uninstall script to switch from winesapOS back to upstream Arch Linux.
+- Add support for Mac drivers on Linux LTS 6.1
 
 ## 3.2.1 - The SteamOS 3.4 Update